% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admin0}
\alias{admin0}
\title{Metadata for administrative level 0 (national level) geographic entities}
\format{
A data frame with variables:
\describe{
\item{continent_code}{continent code}
\item{who_region_code}{WHO region code}
\item{admin0_code}{ISO 3166-1 alpha-2 country code}
\item{admin0_name}{country/territory name}
\item{admin0_sovrn}{name of the territory's sovereign state}
\item{who_status}{country/territory a WHO member state status}
\item{lat}{centroid latitude}
\item{lon}{centroid longitude}
\item{flag_url}{a URL to a flag svg image}
\item{population}{population}
\item{admin1_code_type}{the type of admin1 code that is standard to use for this country (e.g. FIPS, ISO 3166-2)}
\item{admin1_entity_name}{the name used in this country for admin1 divisions (e.g. "state", "province")}
\item{admin2_code_type}{the type of admin2 code that is standard to use for this country}
\item{admin2_entity_name}{the name used in this country for admin1 divisions (e.g. "county", "district")}
}
}
\source{
WHO
}
\usage{
admin0
}
\description{
Metadata for administrative level 0 (national level) geographic entities
}
\keyword{datasets}
