% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admin2}
\alias{admin2}
\title{Metadata for administrative level 2 (second sub-national level) geographic entities (counties, districts, etc.)}
\format{
A data frame with variables:
\describe{
\item{admin0_code}{ISO 3166-1 alpha-2 country code that this entity belongs to}
\item{admin1_code}{admin1 code that this entity belongs to (conforming to the standard indicated in \code{admin1_code_type} in \code{\link{admin0}})}
\item{admin1_name}{admin1 name that this entity belongs to}
\item{admin2_code}{admin2 code (conforming to the standard indicated in \code{admin2_code_type} in \code{\link{admin0}})}
\item{admin2_name}{admin2 name}
\item{lat}{centroid latitude}
\item{lon}{centroid longitude}
\item{population}{population}
\item{map_url}{a URL to an image that shows the location of this entity relative to its country/territory}
}
}
\source{
Currently only data for US counties is contained in this file. Others to come as needed by applications. US metadata comes from \href{https://github.com/CSSEGISandData/COVID-19}{JHU CSSE}.
}
\usage{
admin2
}
\description{
Metadata for administrative level 2 (second sub-national level) geographic entities (counties, districts, etc.)
}
\keyword{datasets}
