% Generated by roxygen2 (4.0.1): do not edit by hand
\name{colourramp}
\alias{.colseq}
\alias{.getval}
\alias{colourramp}
\title{Colour palettes}
\usage{
colourramp(n, ramp = "redblue", start = 0.4, log = F)

.getval(ind, val, predind, smooth)

.colseq(x, tmp, start = 0, stop = 1, log = F, smooth = 0.4)
}
\arguments{
\item{n}{number of colour values to be returned}

\item{ramp}{text specifying what ramp to use}

\item{start}{darkness for darkest colour (0 is black)}

\item{log}{logical, should interval be divided linearly?}

\item{ind}{An index specifying the spacing of the values (0 - 1)}

\item{val}{The values to which a smoothed spline is fitted}

\item{predind}{The location for which the smoothed spline function
is evaluated (0 - 1)}

\item{smooth}{the smoothing as \code{spar} in \code{\link{smooth.spline}}}

\item{x}{Number of colors}

\item{tmp}{Array of rgb values (dim=c(ncolors,3)) along which
the color sequence is interpolated}

\item{stop}{the upper bound for brightness of colours (1 is white)}
}
\description{
Functions to help set up colour palettes for use in plots.
}
\keyword{utilities}

