% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compute_dist}
\alias{compute_dist}
\title{Compute distance between points on the globe}
\usage{
compute_dist(lon, lat, lon0, lat0, Rearth = 6371)
}
\arguments{
\item{lon,lat}{points from which distance is measured (in degree)}

\item{lon0,lat0}{points to which distance is measured (in degree)}

\item{Rearth}{Earth's radius (defaults to 6371km)}
}
\value{
Distance between points in km
}
\description{
This function computes distances between an arbitrary number of points on the globe
}
\examples{
## read in some netcdf data
tas <- readNetCDF(system.file("extdata", "annual_CRUTEMv3_1961-90.nc", package="geoutils"), varname="temp")
lons <- attr(tas, 'lon')
lats <- attr(tas, 'lat')
## compute distance from greenwich
gdist <- compute_dist(lons, lats, 0, 50)

## set up new netcdf object for plotting
gdist <- as.matrix(gdist)
class(gdist) <- 'NetCDF'
attr(gdist, 'lon') <- lons
attr(gdist, 'lat') <- lats

## plot the distances
plot(gdist, type='mean', ti=1)
map(add=T, interior=F)
}
\keyword{utilities}

