% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{continent}
\alias{continent}
\title{Metadata for continents}
\format{
A data frame with variables:
\describe{
\item{continent_code}{continent code}
\item{continent_name}{continent name}
\item{population}{population}
\item{map_url}{a URL to an image that shows the location of this entity relative to its country/territory}
}
}
\usage{
continent
}
\description{
Metadata for continents
}
\keyword{datasets}
