% Generated by roxygen2 (4.0.1): do not edit by hand
\name{doy}
\alias{doy}
\title{Compute the day of the year from date}
\usage{
doy(x)
}
\arguments{
\item{x}{Rdates to compute the day of the year}
}
\description{
Given an .Rdate object, this function computes the day of the year (from 1-366)
}
\keyword{utilities}

