\docType{data}
\name{geocolours}
\alias{.bluewhite}
\alias{.brownwhite}
\alias{.gpcc}
\alias{.greenwhite}
\alias{.hsurf}
\alias{.redwhite}
\alias{.soil}
\alias{.water}
\alias{geocolours}
\title{Colours used in the geoutils package}
\format{matrices with RGB colour values (hex colours for \code{.soil}
        and \code{.water})}
\description{
Provides a few (hidden) colour sequences for use with
colfun
}
\keyword{datasets}

