% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{global}
\alias{global}
\title{Metadata for Global}
\format{
A data frame with variables:
\describe{
\item{global_code}{WHO region code}
\item{global_name}{WHO region name}
\item{population}{population}
}
}
\usage{
global
}
\description{
Metadata for Global
}
\keyword{datasets}
