% Generated by roxygen2 (4.0.1): do not edit by hand
\name{merge.NetCDF}
\alias{merge.NetCDF}
\title{Merge objects of type NetCDF along the time dimension}
\usage{
\method{merge}{NetCDF}(x, y = NULL)
}
\arguments{
\item{x}{input object of class 'NetCDF'}

\item{y}{input object of class 'NetCDF'}
}
\description{
This function provides support to merge objects of type NetCDF.
If overlapping time steps are present, the time step from the
second argument are used. If there is a gap between the objects,
missing values will be infilled with regularly spaced time steps
in the gap.
}
\keyword{utils}

