% Generated by roxygen2 (4.0.1): do not edit by hand
\name{movartrend}
\alias{artrend}
\alias{convertx}
\alias{movartrend}
\alias{trendcov}
\alias{trendresid}
\title{Compute trends on NetCDF objects}
\usage{
movartrend(data, trndln, x = NULL, missthresh = 0.8, robust = FALSE,
  gaps = FALSE, plotable = TRUE, resid = FALSE)

trendcov(data, x = NULL, missthresh = 0.8)

trendresid(data, x = NULL, missthresh = 0.8)

artrend(data, x = NULL, missthresh = 0.8)

convertx(x = NULL, data, dat = data)
}
\arguments{
\item{data}{object of type NetCDF containing the time series}

\item{trndln}{length of trend to be computed (in years if time
attribute is present)}

\item{x}{optional covariate to be used as predictor}

\item{missthresh}{threshold to indicate what fraction of non-missing values is
at least required to run the regression}

\item{robust}{logical, should robust regression be used?}

\item{gaps}{logical, should gaps be allowed for?}

\item{plotable}{logical, should output be rearrange to make it plotable?}

\item{resid}{logical, should residuals of regression be returned?}

\item{x}{object of type NetCDF to be expanded to fit
dimensions of \code{dat}}

\item{data}{object of type NetCDF to be used to match
attributes (e.g. time steps etc.)}

\item{dat}{object to infer dimensions of output}
}
\description{
Compute trends on NetCDF objects where \code{movartrend} computes moving
linear regressions on the time series, \code{trendresid} and \code{trendcov}
compute simple linear regression with the latter returning the covariance
matrix of the standard errors (i.e. for multivariate analyses).

The function \code{convertx} is used to expand an object of type NetCDF
to fit another object of type NetCDF. The standard use case would be to
expand a NetCDF object with ensemble mean global mean temperature to an
object with multiple models and simulations. \code{convertx} is used in
\code{movartrend} to convert a potential predictor for the regression.
}
\examples{
## result should be roughly 1
print(movartrend(t(1:100 + rnorm(100, sd=10)), trndln=100)$trend[1])
}
\keyword{utilities}

