% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nefffun}
\alias{nefffun}
\title{Compute effective degrees of freedom for autocorrelated time series}
\usage{
nefffun(N, acf)
}
\arguments{
\item{N}{number of elements in input}

\item{acf}{autocorrelation coefficients (up to lag 2, higher lags not implemented)}
}
\description{
Givne the number of records and the auto-correlation coefficients, this function
computes the effective degrees of freedom of the time series after
Storch and Zwiers (1999).
}
\keyword{utilities}

