% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_colourbar}
\alias{plot_colourbar}
\title{Plot a colourbar}
\usage{
plot_colourbar(levs, cols = NULL, units = NULL, side = 1, ylab = "",
  labels = NULL, xlab = "", nn = 1, center = F,
  cex.axis = par("cex.axis"), sea.col = NULL, sea.name = "N/A", ...)
}
\arguments{
\item{levs}{vector of shading boundaries or list containing \code{levs},
\code{cols}, \code{units} and \code{sea.col}}

\item{cols}{vector of colours}

\item{units}{character string describing units (optional)}

\item{side}{which side should axis labels be drawn on? \code{side = 2,4}
will produce a vertically aligned colour bar}

\item{xlab,ylab}{axes labels label (alternative to describe the units)}

\item{labels}{alternative labels to be used instead of levs}

\item{nn}{spacing of labels to be drawn, \code{nn = 2} draws every other label only}

\item{center}{logical, should labels refer to colour mid-points?}

\item{cex.axis}{character expansion for axis labelling}

\item{sea.col}{colour for missing values, masked values}

\item{sea.name}{description of missing/masked values}
}
\description{
Function to plot a colourbar with specified colours and leves
or with output from \code{plot.NetCDF}.
}
\keyword{utilities}

