% Generated by roxygen2 (4.0.1): do not edit by hand
\name{select_models}
\alias{select_ensembles}
\alias{select_ensmember}
\alias{select_ensmember2}
\alias{select_models}
\title{Select models and ensemble members}
\usage{
select_models(x, modi = 1:nrow(x))

select_ensembles(x, ensi = 1:ncol(x))

select_ensmember(x, modi = 1:nrow(x))

select_ensmember2(x, ensi = 1:ncol(x))
}
\arguments{
\item{x}{object of class 'NetCDF'}

\item{modi}{indices of models to select (first dimension in x, see details)}

\item{ensi}{indices (or text) of runs to select (second dimension of x)}
}
\description{
Functions to select models \code{select_models} and ensemble members
\code{select_ensembles} from an array of model results. With the corresponding
functions \code{select_ensmember} and \code{select_ensmember2} for backwards
compatibility.
}
\details{
The model selector can be either a vector of names corresponding to the row
names in \code{x}, a logical vector of length \code{nrow(x)}, or a numerical
vector with indices to retain.
}
\keyword{utilities}

