% Generated by roxygen2 (4.0.1): do not edit by hand
\name{select_region}
\alias{select_region}
\title{Select region(s)}
\usage{
select_region(x, regi = 1)
}
\arguments{
\item{x}{input array of class NetCDF}

\item{regi}{index or name of region(s) to select}
}
\description{
Function to select spatial entities (2nd last dimension in NetCDF array).
}
\details{
Named regions to select only work if the NetCDF object contains an attribute
'regions' that contains the region names referred to in \code{regi}.
}
\keyword{utilities}

