\name{select_seas}
\alias{select_seas}
\title{Select seasons from 'NetCDF' object}
\usage{
select_seas(x, seas)
}
\arguments{
  \item{x}{object of class 'NetCDF'}

  \item{seas}{fractional year indicating what season to
  select}
}
\description{
Function to select specified season (by fractional year,
i.e. 0.5 for summer) from an object of class 'NetCDF'.
}
\keyword{utilities}

