% Generated by roxygen2 (4.0.1): do not edit by hand
\name{select_time}
\alias{select_time}
\title{Select specified years}
\usage{
select_time(x, startyear, endyear)
}
\arguments{
\item{x}{object of class 'NetCDF'}

\item{startyear}{first year to be selected}

\item{endyear}{last year to be selected}
}
\description{
Function to select a range of year from object of class 'NetCDF'.
}
\examples{
xx <- t(rnorm(400) + c(-2, 0,3,1))
attr(xx, 'time') <- seq(1901, by=0.25, length=400)
class(xx) <- 'NetCDF'

## plot all seasons with individual seasons
plot(xx, type='ts', seas=T, col=1, lty=1, lwd=1)
## highlight subset of data
plot(select_time(xx, 1961, 1990), lwd=2, add=T, seas=T)
}
\keyword{utilities}

