% Generated by roxygen2 (4.0.1): do not edit by hand
\name{string}
\alias{firstcap}
\alias{simplecap}
\alias{string}
\title{Functions to modify text.}
\usage{
simplecap(x)

firstcap(x)
}
\arguments{
\item{x}{text, vector of type character}
}
\description{
These functions modify text to capitalise the first letter of each
word (\code{simplecap}) or the very first letter (\code{firstcap}).
}
\keyword{manip}

