% Generated by roxygen2 (4.0.1): do not edit by hand
\name{time_average}
\alias{time_average}
\title{Create n-year averages of an NetCDF object}
\usage{
time_average(x, n, offset = NULL, type = "start")
}
\arguments{
\item{x}{array of class 'NetCDF'}

\item{n}{number of time steps to average over}

\item{offset}{offset from which to start (in years)}

\item{type}{Alternative to specifying offset,
one of start (the default), even or end, see details}
}
\description{
This function computes n-year averages from an existing NetCDF time series
read in with \code{\link{readNetCDF}}
}
\details{
The automated way to set the offset works as following.
\code{start} starts the \code{n}-year averages at the beginning
of the time series, \code{end} sets the offset so that the last
average ends in the last year of the time series, \code{even}
sets the offset so that \code{modulo(year, n) == 0} for the first year
in the first average (i.e. for \code{n=5} averages such as 1910-1914).

The time attribute reflects the average over the same block as the average
of the data, that is \code{n}-yearly averages will be centered.
}
\examples{
tas <- readNetCDF(system.file("extdata", "annual_CRUTEMv3_1961-90.nc", package="geoutils"), varname="temp")

## compute 5-yearly averages
tas.avg <- time_average(tas, 5, type='start')

## get a grid point with little missing values
si <- which.max(apply(!is.na(tas), 1, sum))

## plot the data
plot(tas, type='ts', si=si, lwd=1)
plot(tas.avg, type='ts', si=si, col=2, lwd=2, add=T)
}
\keyword{utilities}

