\name{timemerge}
\alias{timemerge}
\title{Merge a list of NetCDF objects along the time dimension}
\usage{
timemerge(x)
}
\arguments{
  \item{x}{list of NetCDF objects to be merged}
}
\description{
Function to merge different NetCDF objects (e.g. historical
and scenario simulation)
}
\keyword{utilities}

