% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{who_regions}
\alias{who_regions}
\title{Metadata for WHO regions}
\format{
A data frame with variables:
\describe{
\item{who_region_code}{WHO region code}
\item{who_region_name}{WHO region name}
\item{population}{population}
\item{map_url}{a URL to an image that shows the location of this entity relative to its country/territory}
}
}
\usage{
who_regions
}
\description{
Metadata for WHO regions
}
\keyword{datasets}
