\name{nefffun}
\alias{nefffun}
\title{Compute effective degrees of freedom for autocorrelated time series}
\usage{
nefffun(N, acf)
}
\arguments{
  \item{N}{number of elements in input}

  \item{acf}{autocorrelation coefficients (up to lag 2,
  higher lags not implemented)}
}
\description{
Compute effective degrees of freedom for autocorrelated
time series
}
\keyword{utilities}

