\name{select_time}
\alias{select_time}
\title{Select specified years}
\usage{
select_time(x, startyear, endyear)
}
\arguments{
  \item{x}{object of class 'NetCDF'}

  \item{startyear}{first year to be selected}

  \item{endyear}{last year to be selected}
}
\description{
Function to select a range of year from object of class
'NetCDF'.
}
\keyword{utilities}

